+++ CHANGELOG.md
@@ -12,10 +12,12 @@
     * Make it possible to include debug bar also into non-HTML responses
 
 # v1.3.1
-## 07/xx/2017
+## 07/19/2017
 
 1. [](#bugfix)
+    * Fix ordering for Linux + International environments [#1574](https://github.com/getgrav/grav/issues/1574)
     * Check if medium thumbnail exists before resetting
+    * Update Travis' auth token
 
 # v1.3.0
 ## 07/16/2017